Split sendmsg feature into setup and usage
Separate sendmsg usage in code files from setting up the package and
config.

This is needed in situations where sendmsg is provided by other roles
which would collide with the setup procedure in this role.

The goal is achieved by introducing a new sendmsg flag for setting up as
opposed to using it. The change is backward compatible by using the same
default for the new flag mon_setup_sendmsg as the existing flag mon_use_sendmsg